+++ farm-erp-ui/src/app/components/harvest/deliveries-table/deliveries-table.component.ts
@@ -15,88 +15,86 @@ import {DatePipe} from "@angular/common";
 })
 export class DeliveriesTableComponent implements OnInit {
 
-  pageSize  = 10;
+  pageSize = 10;
   pageIndex = 0;
   length: any;
-  pageSizeOptions = [5,10,20];
+  pageSizeOptions = [5, 10, 20];
 
   displayedColumns: string[] = [
-    'worker', 'vehicle', 'gross' , 'empty', 'net', 'from', 'where', 'intakeDate', 'update', 'delete'
+    'worker', 'vehicle', 'gross', 'empty', 'net', 'from', 'where', 'intakeDate', 'update', 'delete'
   ];
 
-  dataSource : MatTableDataSource<Deliveries>
-  @Input() diary!: Diary | null
+  dataSource: MatTableDataSource<Deliveries>
   searchCriteria: any;
 
+  @Input() diary!: Diary | null
+
   constructor(
-    private deliveriesService : DeliveriesService,
-    private dialogService     : DialogService,
-    private roleService       : RoleService,
+    private deliveriesService: DeliveriesService,
+    private dialogService: DialogService,
+    private roleService: RoleService,
   ) {
     this.dataSource = new MatTableDataSource<Deliveries>();
   }
 
   ngOnInit(): void {
-    this.deliveriesService.getDeliveriesByDiary(this.diary?.id ?? -1,this.searchCriteria,this.pageIndex,this.pageSize).subscribe(
+    this.deliveriesService.getDeliveriesByDiary(this.diary?.id ?? -1, this.searchCriteria, this.pageIndex, this.pageSize).subscribe(
       deliveries => {
-        console.log(deliveries);
-        if (deliveries.code === 200){
+        if (deliveries.code === 200) {
           this.dataSource.data = deliveries.data.objects[0];
           this.length = deliveries.data.allNumber;
         }
-      }
-    )
+      });
   }
 
   update(deliveries: Deliveries) {
-    if (this.roleService.openPopUpIfNotBoss()){
+    if (this.roleService.openPopUpIfNotBoss()) {
       return
     }
 
-    this.deliveriesService.openCreateDialog(this.diary?.id ?? -1,false, deliveries).subscribe(
+    this.deliveriesService.openCreateDialog(this.diary?.id ?? -1, false, deliveries).subscribe(
       data => {
-        console.log(data);
-        if (data){
+        if (data) {
           const datepipe: DatePipe = new DatePipe('en-US');
 
           data.intakeDate = String(datepipe.transform(data.intakeDate, 'yyyy-MM-dd HH:mm:ss') ?? '');
           this.deliveriesService.updateDelivery(data).subscribe(
             data => {
-              if(data.code === 200 || data.code === 502){
+              if (data.code === 200) {
                 this.dialogService.openSuccessDialog();
-              }else{
+              } else if (data.code === 1501) {
+                this.dialogService.openStorageIsFullDialog();
+              } else {
                 this.dialogService.openUnexpectedErrorDialog();
               }
               this.ngOnInit();
-            }
-          )
+            });
         }
-      }
-    )
-
-
+      });
   }
 
   delete(deliveries: Deliveries) {
-    if (this.roleService.openPopUpIfNotBoss()){
+    if (this.roleService.openPopUpIfNotBoss()) {
       return
     }
 
-    this.deliveriesService.deleteDelivery(deliveries.id).subscribe(
-      data => {
-        if(data.code === 200){
-          this.dialogService.openSuccessDialog();
-        }else{
-          this.dialogService.openUnexpectedErrorDialog();
-        }
-        this.ngOnInit();
-      }
-    )
+    this.dialogService.openConfirmDialog("Biztosan szeretné törölni?", "warn").afterClosed().subscribe(res => {
+      if (res)
+        this.deliveriesService.deleteDelivery(deliveries.id).subscribe(
+          data => {
+            if (data.code === 200) {
+              this.dialogService.openSuccessDialog();
+            } else {
+              this.dialogService.openUnexpectedErrorDialog();
+            }
+            this.ngOnInit();
+          });
+    });
   }
 
   pageEvent(page: PageEvent) {
     this.pageIndex = page.pageIndex;
-    this.pageSize  = page.pageSize;
+    this.pageSize = page.pageSize;
 
     this.ngOnInit();
   }
@@ -107,8 +105,8 @@ export class DeliveriesTableComponent implements OnInit {
 
   clear() {
     this.searchCriteria = "";
-    this.pageIndex      = 0;
-    this.pageSize       = 10;
+    this.pageIndex = 0;
+    this.pageSize = 10;
 
     this.ngOnInit();
   }
@@ -116,25 +114,22 @@ export class DeliveriesTableComponent implements OnInit {
   createDeliver() {
     this.deliveriesService.openCreateDialog(this.diary?.id ?? -1).subscribe(
       data => {
-        console.log(data);
-        if (data){
+        if (data) {
           const datepipe: DatePipe = new DatePipe('en-US');
 
           data.intakeDate = String(datepipe.transform(data.intakeDate, 'yyyy-MM-dd HH:mm:ss') ?? '');
           this.deliveriesService.createNewDelivery(data).subscribe(
             data => {
-              if(data.code === 200 || data.code === 502){
+              if (data.code === 200) {
                 this.dialogService.openSuccessDialog();
-              }else{
+              } else if (data.code === 1501) {
+                this.dialogService.openStorageIsFullDialog();
+              } else {
                 this.dialogService.openUnexpectedErrorDialog();
               }
               this.ngOnInit();
-            }
-          )
+            });
         }
-      }
-    )
+      });
   }
-
-
 }
